Added all in-game flowers to designer
Add every flower in-game to the designer

Added every flower in-game to the designer (Custom captured and grown assets of the flowers).

For possible compatibility issues I maintained the names of the original flower names in flower.ts and the rest are rewritten in a different format.

Enabled weeds that were commented out.